+++ notification_configuration_integration_test.go
@@ -852,3 +852,388 @@ func TestNotificationConfigurationVerify_forTeams(t *testing.T) {
 		assert.Equal(t, err, ErrInvalidNotificationConfigID)
 	})
 }
+
+func TestNotificationConfigurationList_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	ncTest1, ncTestCleanup1 := createProjectNotificationConfiguration(t, client, pTest, nil)
+	t.Cleanup(ncTestCleanup1)
+	ncTest2, ncTestCleanup2 := createProjectNotificationConfiguration(t, client, pTest, nil)
+	t.Cleanup(ncTestCleanup2)
+
+	t.Run("with a valid project", func(t *testing.T) {
+		ncl, err := client.NotificationConfigurations.List(
+			ctx,
+			pTest.ID,
+			&NotificationConfigurationListOptions{
+				SubscribableChoice: &NotificationConfigurationSubscribableChoice{
+					Project: pTest,
+				},
+			},
+		)
+		require.NoError(t, err)
+		assert.Contains(t, ncl.Items, ncTest1)
+		assert.Contains(t, ncl.Items, ncTest2)
+	})
+
+	t.Run("without a valid project", func(t *testing.T) {
+		ncl, err := client.NotificationConfigurations.List(
+			ctx,
+			badIdentifier,
+			&NotificationConfigurationListOptions{
+				SubscribableChoice: &NotificationConfigurationSubscribableChoice{
+					Project: pTest,
+				},
+			},
+		)
+		assert.Nil(t, ncl)
+		assert.EqualError(t, err, ErrInvalidProjectID.Error())
+	})
+}
+
+func TestNotificationConfigurationCreate_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	// Create user to use when testing email destination type
+	orgMemberTest, orgMemberTestCleanup := createOrganizationMembership(t, client, orgTest)
+	t.Cleanup(orgMemberTestCleanup)
+
+	t.Run("with all required values", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeGeneric),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			Token:              String(randomString(t)),
+			URL:                String("http://example.com"),
+			Triggers:           []NotificationTriggerType{NotificationTriggerCreated},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+
+		require.NoError(t, err)
+		require.NotNil(t, nc)
+	})
+
+	t.Run("without a required value", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeGeneric),
+			Enabled:            Bool(false),
+			Token:              String(randomString(t)),
+			URL:                String("http://example.com"),
+			Triggers:           []NotificationTriggerType{NotificationTriggerCreated},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+
+		assert.Nil(t, nc)
+		assert.EqualError(t, err, ErrRequiredName.Error())
+	})
+
+	t.Run("without a required value URL when destination type is generic", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeGeneric),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			Token:              String(randomString(t)),
+			Triggers:           []NotificationTriggerType{NotificationTriggerCreated},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		assert.Nil(t, nc)
+		assert.Equal(t, err, ErrRequiredURL)
+	})
+
+	t.Run("without a required value URL when destination type is slack", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeSlack),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			Triggers:           []NotificationTriggerType{NotificationTriggerCreated},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		assert.Nil(t, nc)
+		assert.Equal(t, err, ErrRequiredURL)
+	})
+
+	t.Run("without a required value URL when destination type is MS Teams", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeMicrosoftTeams),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			Triggers:           []NotificationTriggerType{NotificationTriggerCreated},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		assert.Nil(t, nc)
+		assert.Equal(t, err, ErrRequiredURL)
+	})
+
+	t.Run("without a valid project", func(t *testing.T) {
+		nc, err := client.NotificationConfigurations.Create(ctx, badIdentifier, NotificationConfigurationCreateOptions{
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{
+				Project: pTest,
+			},
+		})
+		assert.Nil(t, nc)
+		assert.EqualError(t, err, ErrInvalidProjectID.Error())
+	})
+
+	t.Run("with an invalid notification trigger", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeGeneric),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			Token:              String(randomString(t)),
+			URL:                String("http://example.com"),
+			Triggers:           []NotificationTriggerType{"the beacons of gondor are lit"},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		nc, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		assert.Nil(t, nc)
+		assert.EqualError(t, err, ErrInvalidNotificationTrigger.Error())
+	})
+
+	t.Run("with email users when destination type is email", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeEmail),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			EmailUsers:         []*User{orgMemberTest.User},
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		_, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		require.NoError(t, err)
+	})
+
+	t.Run("without email users when destination type is email", func(t *testing.T) {
+		options := NotificationConfigurationCreateOptions{
+			DestinationType:    NotificationDestination(NotificationDestinationTypeEmail),
+			Enabled:            Bool(false),
+			Name:               String(randomString(t)),
+			SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+		}
+
+		_, err := client.NotificationConfigurations.Create(ctx, pTest.ID, options)
+		require.NoError(t, err)
+	})
+}
+
+func TestNotificationConfigurationRead_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	ncTest, ncTestCleanup := createProjectNotificationConfiguration(t, client, pTest, nil)
+	t.Cleanup(ncTestCleanup)
+
+	t.Run("with a valid ID", func(t *testing.T) {
+		nc, err := client.NotificationConfigurations.Read(ctx, ncTest.ID)
+		require.NoError(t, err)
+		assert.Equal(t, ncTest.ID, nc.ID)
+	})
+
+	t.Run("when the notification configuration does not exist", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Read(ctx, "nonexisting")
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("when the notification configuration ID is invalid", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Read(ctx, badIdentifier)
+		assert.Equal(t, err, ErrInvalidNotificationConfigID)
+	})
+}
+
+func TestNotificationConfigurationUpdate_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	ncTest, ncTestCleanup := createProjectNotificationConfiguration(t, client, pTest, nil)
+	t.Cleanup(ncTestCleanup)
+
+	// Create users to use when testing email destination type
+	orgMemberTest1, orgMemberTest1Cleanup := createOrganizationMembership(t, client, orgTest)
+	defer orgMemberTest1Cleanup()
+	orgMemberTest2, orgMemberTest2Cleanup := createOrganizationMembership(t, client, orgTest)
+	defer orgMemberTest2Cleanup()
+
+	orgMemberTest1.User = &User{ID: orgMemberTest1.User.ID}
+	orgMemberTest2.User = &User{ID: orgMemberTest2.User.ID}
+
+	emailOptions := &NotificationConfigurationCreateOptions{
+		DestinationType:    NotificationDestination(NotificationDestinationTypeEmail),
+		Enabled:            Bool(false),
+		Name:               String(randomString(t)),
+		EmailUsers:         []*User{orgMemberTest1.User},
+		SubscribableChoice: &NotificationConfigurationSubscribableChoice{Project: pTest},
+	}
+	ncEmailTest, ncEmailTestCleanup := createProjectNotificationConfiguration(t, client, pTest, emailOptions)
+	t.Cleanup(ncEmailTestCleanup)
+
+	t.Run("with options", func(t *testing.T) {
+		options := NotificationConfigurationUpdateOptions{
+			Enabled: Bool(true),
+			Name:    String("newName"),
+		}
+
+		nc, err := client.NotificationConfigurations.Update(ctx, ncTest.ID, options)
+		require.NoError(t, err)
+		assert.Equal(t, nc.Enabled, true)
+		assert.Equal(t, nc.Name, "newName")
+	})
+
+	t.Run("with invalid notification trigger", func(t *testing.T) {
+		options := NotificationConfigurationUpdateOptions{
+			Triggers: []NotificationTriggerType{"fly you fools!"},
+		}
+
+		nc, err := client.NotificationConfigurations.Update(ctx, ncTest.ID, options)
+		assert.Nil(t, nc)
+		assert.EqualError(t, err, ErrInvalidNotificationTrigger.Error())
+	})
+
+	t.Run("with email users when destination type is email", func(t *testing.T) {
+		options := NotificationConfigurationUpdateOptions{
+			Enabled:    Bool(true),
+			Name:       String("newName"),
+			EmailUsers: []*User{orgMemberTest1.User, orgMemberTest2.User},
+		}
+
+		nc, err := client.NotificationConfigurations.Update(ctx, ncEmailTest.ID, options)
+		require.NoError(t, err)
+		assert.Equal(t, nc.Enabled, true)
+		assert.Equal(t, nc.Name, "newName")
+		assert.Contains(t, nc.EmailUsers, orgMemberTest1.User)
+		assert.Contains(t, nc.EmailUsers, orgMemberTest2.User)
+	})
+
+	t.Run("without email users when destination type is email", func(t *testing.T) {
+		options := NotificationConfigurationUpdateOptions{
+			Enabled: Bool(true),
+			Name:    String("newName"),
+		}
+
+		nc, err := client.NotificationConfigurations.Update(ctx, ncEmailTest.ID, options)
+		require.NoError(t, err)
+		assert.Equal(t, nc.Enabled, true)
+		assert.Equal(t, nc.Name, "newName")
+		assert.Empty(t, nc.EmailUsers)
+	})
+
+	t.Run("without options", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Update(ctx, ncTest.ID, NotificationConfigurationUpdateOptions{})
+		require.NoError(t, err)
+	})
+
+	t.Run("when the notification configuration does not exist", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Update(ctx, "nonexisting", NotificationConfigurationUpdateOptions{})
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("when the notification configuration ID is invalid", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Update(ctx, badIdentifier, NotificationConfigurationUpdateOptions{})
+		assert.Equal(t, err, ErrInvalidNotificationConfigID)
+	})
+}
+
+func TestNotificationConfigurationDelete_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	ncTest, _ := createProjectNotificationConfiguration(t, client, pTest, nil)
+
+	t.Run("with a valid ID", func(t *testing.T) {
+		err := client.NotificationConfigurations.Delete(ctx, ncTest.ID)
+		require.NoError(t, err)
+
+		_, err = client.NotificationConfigurations.Read(ctx, ncTest.ID)
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("when the notification configuration does not exist", func(t *testing.T) {
+		err := client.NotificationConfigurations.Delete(ctx, "nonexisting")
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("when the notification configuration ID is invalid", func(t *testing.T) {
+		err := client.NotificationConfigurations.Delete(ctx, badIdentifier)
+		assert.Equal(t, err, ErrInvalidNotificationConfigID)
+	})
+}
+
+func TestNotificationConfigurationVerify_forProjects(t *testing.T) {
+	t.Parallel()
+	skipUnlessBeta(t)
+	client := testClient(t)
+	ctx := context.Background()
+
+	orgTest, orgTestCleanup := createOrganization(t, client)
+	t.Cleanup(orgTestCleanup)
+
+	pTest, pTestCleanup := createProject(t, client, orgTest)
+	t.Cleanup(pTestCleanup)
+
+	ncTest, ncTestCleanup := createProjectNotificationConfiguration(t, client, pTest, nil)
+	t.Cleanup(ncTestCleanup)
+
+	t.Run("with a valid ID", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Verify(ctx, ncTest.ID)
+		require.NoError(t, err)
+	})
+
+	t.Run("when the notification configuration does not exists", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Verify(ctx, "nonexisting")
+		assert.Equal(t, err, ErrResourceNotFound)
+	})
+
+	t.Run("when the notification configuration ID is invalid", func(t *testing.T) {
+		_, err := client.NotificationConfigurations.Verify(ctx, badIdentifier)
+		assert.Equal(t, err, ErrInvalidNotificationConfigID)
+	})
+}